--- segmentstore/storage/impl/src/test/java/io/pravega/segmentstore/storage/impl/bookkeeper/SecureBookKeeperLogTests.java
@@ -1,20 +1,0 @@
-/**
- * Copyright (c) 2017 Dell Inc., or its subsidiaries. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- */
-package io.pravega.segmentstore.storage.impl.bookkeeper;
-
-import org.junit.BeforeClass;
-
-public class SecureBookKeeperLogTests extends BookKeeperLogTests {
-    @BeforeClass
-    public static void startUp() throws Exception {
-        SECURE_BK.set(true);
-        setUpBookKeeper();
-    }
-}
